Refactor: Separate navigation for main app and news page
- Created a new MainNavbar component for the main application navigation.
- Added a 'Health Magazine' link to the MainNavbar.
- Renamed the existing Navbar to NewsHeader to serve as the header for the /news page.
- Updated all relevant pages to use the correct navigation component.
