+++ client/src/components/MainNavbar.css
@@ -1,0 +1,150 @@
+/* Reset‌های لازم و یکدست‌سازی رندر متن */
+html,body{margin:0;padding:0}
+html{ -webkit-text-size-adjust:100%; text-size-adjust:100% }
+*{ box-sizing:border-box }
+:root{
+  --primary:#1565c0;
+  --danger:#d32f2f;
+  --text-on-primary:#fff;
+}
+body{
+  font-family:'Vazirmatn', Tahoma, sans-serif;
+  background:#f5f7fb;
+  color:#1f2937;
+  line-height:1.6;
+}
+/* کمک به یکدست شدن رندر فارسی بین مرورگرها */
+body, .navbar{
+  -webkit-font-smoothing:antialiased; /* کروم/سافاری */
+  -moz-osx-font-smoothing:grayscale;  /* فایرفاکس روی مک */
+  font-kerning:normal;
+  font-feature-settings:"kern","liga","clig","calt";
+}
+
+/* ناوبَر */
+.navbar{
+  display:flex;
+  justify-content:space-between;
+  align-items:center;
+  gap:1rem;
+  background-color:var(--primary);
+  color:var(--text-on-primary);
+  padding:.875rem 1.25rem;
+  border-bottom:1px solid rgba(255,255,255,.12);
+  position: relative;
+  top: 0;
+  z-index: 1000;
+}
+/* Desktop First Layout */
+.navbar-left, .navbar-center, .navbar-right {
+  display: flex;
+  align-items: center;
+}
+.navbar-left { flex: 1; justify-content: flex-start; }
+.navbar-center { flex: 2; justify-content: center; }
+.navbar-right { flex: 1; justify-content: flex-end; }
+
+.navbar-brand a {
+  color: #fff; text-decoration: none;
+  font-weight: 700; font-size: 1.375rem; white-space: nowrap;
+}
+
+.navbar-links {
+  display: flex; align-items: center; gap: .75rem;
+}
+.navbar-links a {
+  color: #fff; text-decoration: none;
+  padding: .4rem .6rem; border-radius: .5rem;
+  font-weight: 500; line-height: 1.4; white-space: nowrap;
+}
+.navbar-links a:hover { background-color: rgba(255,255,255,.12); }
+
+.navbar-profile { display: flex; align-items: center; gap: .5rem; }
+.navbar-profile span { font-size: 1.25rem; cursor: pointer; line-height: 1; }
+
+.btn {
+  display: inline-flex; align-items: center; justify-content: center;
+  gap: .4rem; padding: .5rem .9rem; font-size: 1rem;
+  line-height: 1.5; border-radius: .5rem; cursor: pointer;
+  border: 1px solid #fff; background: transparent; color: #fff;
+  white-space: nowrap; text-decoration: none;
+}
+.btn:focus { outline: 2px solid rgba(255,255,255,.6); outline-offset: 2px; }
+
+/* --- Responsive Profile Links --- */
+.mobile-only-profile { display: none; } /* Hidden on desktop */
+.desktop-only-profile { display: inline-flex; } /* Visible on desktop */
+
+/* --- Toggler --- */
+.navbar-toggler {
+  display: none; /* Hidden on desktop */
+  font-size: 1.5rem; background: transparent;
+  border: none; color: white; cursor: pointer;
+}
+
+.admin-link {
+  color: #fdd835 !important; /* A distinct yellow color for visibility */
+  font-weight: 700;
+}
+
+/* --- Mobile View --- */
+.menu-backdrop {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.5);
+  z-index: 1999; /* Below the menu (2000), above everything else */
+}
+
+@media (max-width: 768px) {
+  .navbar-center {
+    position: fixed;
+    top: 0;
+    right: 0;
+    width: 280px; /* A fixed width for the sidebar */
+    height: 100%;
+    background-color: var(--primary);
+    padding: 1.5rem 1rem;
+    z-index: 2000;
+    display: block;
+    transform: translateX(100%);
+    transition: transform 0.3s ease-in-out;
+  }
+
+  .navbar-center.active {
+    transform: translateX(0);
+  }
+
+  .navbar-links {
+    flex-direction: column;
+    align-items: flex-start; /* Align links to the start (right in RTL) */
+    width: 100%;
+  }
+
+  .navbar-links a {
+    text-align: center;
+    padding: .75rem 1rem;
+    border-bottom: 1px solid rgba(255, 255, 255, 0.1);
+  }
+
+  .navbar-links a:last-child {
+    border-bottom: none;
+  }
+
+  .desktop-only-profile {
+    display: none;
+  }
+
+  .mobile-only-profile {
+    display: flex;
+    justify-content: center;
+    background-color: rgba(0,0,0,0.1);
+    margin-top: 0.5rem;
+  }
+
+  .navbar-toggler {
+    display: block;
+  }
+}
