+++ server/db.json
@@ -224,7 +224,7 @@
         "category": "تربیتی",
         "imageUrl": "/uploads/sample-image-3.jpg",
         "createdAt": "2023-10-24T09:00:00Z"
-    }
+    },
         {
         "id": 4,
         "title": "سرماخوردگی در نوزادان: علائم و درمان",
